Reworked animation resources
Animations are now stored with a reference to a 'frame set',
so that frame sets can be used to assign frames. All the animation
contains now is the sprite sheet, the frame set informs what frames
are extracted and played.

This means it's much easier to sync up animations of 2 entities, like
when you need to sync up the animations of equipment and the entity wearing
that equipment.

Also introduced a new system which tracks positions - i.e. it matches up
an entities position, and allows one entities position to just follow
another entity's position (also useful for rendering equipment!)